+++ src/index.ts
@@ -57,6 +57,7 @@ export { default as Avatar } from '@/package/components/avatar.vue';
 export { default as DataTable } from '@/package/components/data-table.vue';
 export { default as Slider } from '@/package/components/slider.vue';
 export { default as Stepper } from '@/package/components/stepper.vue';
+export { default as Timeline } from '@/package/components/timeline.vue';
 export { default as Rating } from '@/package/components/rating.vue';
 export { default as Tree } from '@/package/components/tree.vue';
 export { default as TreeSelect } from '@/package/components/tree-select.vue';
@@ -64,6 +65,7 @@ export { default as Breadcrumbs } from '@/package/components/breadcrumbs.vue';
 export { default as VirtualScroller } from '@/package/components/virtual-scroller.vue';
 export type { DataTableColumn } from '@/package/components/data-table.vue';
 export type { TreeItem, TreeValue } from '@/package/components/tree.vue';
+export type { TimelineItem } from '@/package/components/timeline.vue';
 // config
 export { default as VueForge, setTheme, updateTheme, getTheme } from '@/package/config/index';
 export type { ThemeTokens, ThemeOptions, ThemePreset } from '@/package/config/theme-core';